Move SvgIcon to the components package
