+++ packages/components/Notification.js
@@ -7,7 +7,7 @@ import { __ } from "@wordpress/i18n";
 /* Internal dependencies */
 import Paper from "./Paper";
 import { colors, breakpoints } from "./style-guide";
-import SvgIcon from "../yoast-components/composites/Plugin/Shared/components/SvgIcon";
+import SvgIcon from "./SvgIcon";
 
 const NotificationContainer = styled.div`
 	display: flex;
