+++ packages/components/SvgIcon.js
