+++ packages/components/style-guide/index.js
@@ -1,3 +1,9 @@
-export colors from "colors";
-export breakpoints from "responsive-breakpoints";
-export { rgba } from "helpers";
+import colors from "./colors";
+import breakpoints from "./responsive-breakpoints";
+import { rgba } from "./helpers";
+
+export {
+	colors,
+	breakpoints,
+	rgba
+};
