+++ packages/yoast-components/composites/AlgoliaSearch/SearchResultDetail.js
@@ -6,7 +6,7 @@ import { defineMessages, injectIntl, intlShape } from "react-intl";
 import ArticleContent from "./ArticleContent";
 import { YoastButton } from "../Plugin/Shared/components/YoastButton";
 import { YoastLinkButton } from "../Plugin/Shared/components/YoastLinkButton";
-import SvgIcon from "../Plugin/Shared/components/SvgIcon";
+import { SvgIcon } from "@yoast/components";
 import { makeOutboundLink } from "../../utils/makeOutboundLink";
 import { colors, breakpoints } from "@yoast/components/style-guide";
