+++ packages/yoast-components/composites/Plugin/Shared/components/HelpCenterButton.js
@@ -4,7 +4,7 @@ import styled from "styled-components";
 
 import { colors } from "@yoast/components/style-guide";
 import { YoastButton } from "./YoastButton";
-import SvgIcon from "../../Shared/components/SvgIcon";
+import { SvgIcon } from "@yoast/components";
 
 /**
  * Returns an icon button that can optionally contain text.
