+++ packages/yoast-components/composites/Plugin/Shared/components/IconButtonToggle.js
@@ -3,7 +3,7 @@ import styled from "styled-components";
 import PropTypes from "prop-types";
 
 import { colors, rgba } from "@yoast/components/style-guide";
-import SvgIcon from "./SvgIcon";
+import { SvgIcon } from "@yoast/components";
 
 const IconButtonBase = styled.button`
 	box-sizing: border-box;
