+++ packages/yoast-components/composites/Plugin/Shared/components/IconLabelledButton.js
@@ -5,7 +5,7 @@ import flow from "lodash/flow";
 import omit from "lodash/omit";
 
 import { colors } from "@yoast/components/style-guide";
-import SvgIcon from "./SvgIcon";
+import { SvgIcon } from "@yoast/components";
 // Note: the order of these imports matters for the styles order in the markup.
 import { addActiveStyle, addFocusStyle, addHoverStyle } from "./Button";
