+++ packages/yoast-components/composites/Plugin/SnippetEditor/components/ModeSwitcher.js
@@ -8,7 +8,7 @@ import { Button } from "../../Shared/components/Button";
 import { colors } from "@yoast/components/style-guide";
 import { MODE_DESKTOP, MODE_MOBILE, MODES } from "../../SnippetPreview/constants";
 import ScreenReaderText from "../../../../a11y/ScreenReaderText";
-import SvgIcon from "../../Shared/components/SvgIcon";
+import { SvgIcon } from "@yoast/components";
 import PropTypes from "prop-types";
 
 /**
